projekts pabegts ar skaistiem laciem fona

+++ JK_saldumi_31_01/Form1.cs
@@ -24,7 +24,7 @@ namespace JK_saldumi_31_01
 
         private void aprekinat_Click(object sender, EventArgs e)
         {
-            double naudas_daudzums_eur = Convert.ToInt32(this.naudas_ievade.Text);
+            double naudas_daudzums_eur = Convert.ToDouble(this.naudas_ievade.Text);
             double rezultats_eur = 0;
             rezultats_eur = naudas_daudzums_eur /9.88;
             rezultats_eur = Math.Round(rezultats_eur, 2);
